+++ libs/fsm/tests/event_queue.c
@@ -150,7 +150,7 @@ int main(void) {
             AM_POOL_BLOCK_ALIGNMENT(AM_ALIGNOF(am_event_t))
         );
         AM_ASSERT(1 == am_event_get_pool_nblocks(/*index=*/0));
-        AM_ASSERT(1 == am_event_get_pool_nfree(/*index=*/0));
+        AM_ASSERT(1 == am_event_pool_get_nfree(/*index=*/0));
     }
 
     fsmq_ctor(fsmq_log);
@@ -174,7 +174,7 @@ int main(void) {
     am_fsm_dtor(am_fsmq);
 
     /* make sure there is no memory leak */
-    AM_ASSERT(1 == am_event_get_pool_nfree(/*index=*/0));
+    AM_ASSERT(1 == am_event_pool_get_nfree(/*index=*/0));
 
     return 0;
 }