+++ libs/hsm/tests/defer.c
@@ -178,7 +178,7 @@ static void test_defer(void) {
             AM_POOL_BLOCK_ALIGNMENT(AM_ALIGNOF(am_event_t))
         );
         AM_ASSERT(2 == am_event_get_pool_nblocks(/*index=*/0));
-        AM_ASSERT(2 == am_event_get_pool_nfree(/*index=*/0));
+        AM_ASSERT(2 == am_event_pool_get_nfree(/*index=*/0));
     }
 
     defer_ctor(defer_hsm_log);
@@ -206,7 +206,7 @@ static void test_defer(void) {
     }
 
     /* make sure there is no memory leak */
-    AM_ASSERT(2 == am_event_get_pool_nfree(/*index=*/0));
+    AM_ASSERT(2 == am_event_pool_get_nfree(/*index=*/0));
 }
 
 int main(void) {