+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Rename `am_event_queue_get_nbusy()` to `am_event_queue_get_nbusy_unsafe()`.
 - Rename `am_event_add_pool()` to  `am_event_pool_add()`.
 - Rename `am_event_get_pool_nfree_min()` to `am_event_pool_get_nfree_min()`.
+- Rename `am_event_get_pool_nfree()` to `am_event_pool_get_nfree()`.
 
 ### Added